% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_election.R
\name{print.rs_election}
\alias{print.rs_election}
\title{Generic to print text summary of the rs_election object}
\usage{
\method{print}{rs_election}(x, ...)
}
\arguments{
\item{x}{rs_election}

\item{...}{Additional arguments}
}
\value{
A text summary of the rs_election object - Returns rs_election invisibly
}
\description{
Generic to print text summary of the rs_election object
}
