% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_election.R
\name{rs_adjustment}
\alias{rs_adjustment}
\title{Adjust for votes not cast at an ordinary polling place}
\usage{
rs_adjustment(rs_election_results, rs_polling_places)
}
\arguments{
\item{rs_election_results}{An rs_election_results object}

\item{rs_polling_places}{An rs_polling_places object}
}
\value{
An rs_adjustment object
}
\description{
Adjust for votes not cast at an ordinary polling place
}
