% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_boundary_set.R
\name{rs_boundary_set}
\alias{rs_boundary_set}
\title{Create an rs_boundary_set}
\usage{
rs_boundary_set(sfc, names, to_crs = NULL)
}
\arguments{
\item{sfc}{Simple feature collection}

\item{names}{Name of the column containing electorate/district names}

\item{to_crs}{A CRS to convert the sfc to for the purposes of st_union()}
}
\value{
A simple feature collection with two columns, district and geometry
}
\description{
Create an rs_boundary_set
}
