% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_election.R
\name{rs_election}
\alias{rs_election}
\title{Create an rs_election object with voronoi polygons for polling booths}
\usage{
rs_election(name, boundaries, polling_places, election_result_list, crs)
}
\arguments{
\item{name}{The name of the election}

\item{boundaries}{An rs_boundary_set with the electorate boundaries}

\item{polling_places}{An rs_polling_places object with each polling place used}

\item{election_result_list}{An list that contains one or more rs_election_results}

\item{crs}{The CRS that the voronoi polygons will be made in. Typically a projected CRS for the relevant area.}
}
\value{
An rs_election object
}
\description{
Create an rs_election object with voronoi polygons for polling booths
}
