% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_election_as_sf.R
\name{rs_election_as_sf}
\alias{rs_election_as_sf}
\title{Adds colors based on the imported color palette}
\usage{
rs_election_as_sf(
  election,
  result_name,
  polcolpal,
  adjustments = TRUE,
  mode = "plurality",
  single_group_code = NA
)
}
\arguments{
\item{election}{An rs_election or rs_redistribution object}

\item{result_name}{The name of the re_election_results object in the election parameter}

\item{polcolpal}{Political color palette created using rs_polcolpal_create}

\item{adjustments}{Whether to calculate the color based on the on-the-day vote or after adjustments are made}

\item{mode}{Either "plurality" or "single color". "Plurality" gives the largest vote at the booth, "single color" plots the vote share of the group specified in single_group_code}

\item{single_group_code}{If mode == "single color", the group_code to color with}
}
\value{
An rs_election_as_sf object with a col_code column
}
\description{
Adds colors based on the imported color palette
}
