% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_election_results.R
\name{rs_election_results}
\alias{rs_election_results}
\title{Creates an rs_election_results object}
\usage{
rs_election_results(name, df, pp_id, group_code, votes)
}
\arguments{
\item{name}{The name of the result (Primary Vote, Two Party Preferred, Two Candidate Preferred etc.)}

\item{df}{The data frame that lists the results by polling booth}

\item{pp_id}{The column containing the polling place identifier}

\item{group_code}{The column containing the variable to split the results by (Typically, this refers to political parties or party codes)}

\item{votes}{The column containing the votes that each group recieved at each polling place}
}
\value{
An rs_election_results object
}
\description{
Creates an rs_election_results object
}
