% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_plot.R
\name{rs_plot}
\alias{rs_plot}
\title{Plots an rs_election or rs_redistribution object with a legend}
\usage{
rs_plot(x, pdf_file_name = NA)
}
\arguments{
\item{x}{An rs_election_as_sf object}

\item{pdf_file_name}{Name of the pdf file. If NA, then plots to viewer.}
}
\value{
Plots the object or returns a pdf file
}
\description{
Plots an rs_election or rs_redistribution object with a legend
}
