% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_polcolpal_create.R
\name{rs_polcolpal_create}
\alias{rs_polcolpal_create}
\title{Creates a color palette for each group code}
\usage{
rs_polcolpal_create(df, breaks)
}
\arguments{
\item{df}{A data frame with two columns. The first lists each group code, the second lists the base color for each group code.}

\item{breaks}{A vector of 10 numbers in ascending order that specifies when the strength of the color should change on the map. Anything below the lowest break will be colored white.}
}
\value{
An rs_polcolpal object
}
\description{
Creates a color palette for each group code
}
