% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_polling_places.R
\name{rs_polling_places}
\alias{rs_polling_places}
\title{Creates an rs_polling_places object}
\usage{
rs_polling_places(
  df,
  pp_id,
  dist_names,
  longitude = "longitude",
  latitude = "latitude",
  crs = 4326
)
}
\arguments{
\item{df}{A data frame of polling places}

\item{pp_id}{The column name of the polling place identifier}

\item{dist_names}{The column name of the district names}

\item{longitude}{The column name of the longitude}

\item{latitude}{The column name of the latitude}

\item{crs}{The CRS of the coordinates (Defaults to 4326)}
}
\value{
An rs_polling_places object
}
\description{
Creates an rs_polling_places object
}
