% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_redistribution.R
\name{rs_redistribution}
\alias{rs_redistribution}
\title{Creates an rs_redistribution object}
\usage{
rs_redistribution(election, new_boundary_set, slivers_rm = TRUE)
}
\arguments{
\item{election}{An rs_election object}

\item{new_boundary_set}{An rs_boundary_set object - representing the new boundaries}

\item{slivers_rm}{If true, any rs_boundary_set element containing less than 2\% of another polling place will have that polling place removed from the calculations for the redistribution.}
}
\value{
An rs_redistribution object
}
\description{
Creates an rs_redistribution object
}
