% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_election.R
\name{summary.rs_election}
\alias{summary.rs_election}
\title{Summaries of each of the results in an rs_election object}
\usage{
\method{summary}{rs_election}(object, ...)
}
\arguments{
\item{object}{An rs_election object}

\item{...}{Additional arguments}
}
\value{
A list of the summarized results, with each list item referring to a different result. The results themselves are summarized by district.
}
\description{
Summaries of each of the results in an rs_election object
}
