% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_redistribution.R
\name{summary.rs_redistribution}
\alias{summary.rs_redistribution}
\title{Generic that summarizes an rs_redistribution}
\usage{
\method{summary}{rs_redistribution}(object, ...)
}
\arguments{
\item{object}{An rs_redistribution object}

\item{...}{Additional arguments}
}
\value{
A list of summarized results for the new boundaries
}
\description{
Generic that summarizes an rs_redistribution
}
